requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==25.1.0
    # via autogit (pyproject.toml)
anyio==4.11.0
    # via httpx
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via autogit (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via autogit (pyproject.toml)
idna==3.11
    # via
    #   anyio
    #   httpx
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
pygments==2.19.2
    # via rich
rich==14.2.0
    # via autogit (pyproject.toml)
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
typing-extensions==4.15.0
    # via anyio
